Event_Scheduler/requirements.txt
Flask-SQLAlchemy
Celery